program that prints the size of various types on the computer it is compiled and run on.

+++ 0x00-hello_world/6-size.c
@@ -9,7 +9,7 @@ int main(void)
 	int intType;
 	char charType;
 	long longType;
-	long longlongType;
+	double longlongType;
 	float floatType;
 
 	printf("Size of a char: %zu bytes\n", sizeof(charType));
